UC-17 Changes of Liter & Milliliter inputs from int to float.

+++ src/main/java/LiterMilliliter.java
@@ -1,13 +1,13 @@
 public class LiterMilliliter {
 
 
-    public int convertLiterToMilliliter(int liter) {
+    public int convertLiterToMilliliter(float liter) {
         System.out.println();
-        int milliliter = liter * 1000;
+        int milliliter = (int) (liter * 1000);
         return milliliter;
     }
 
-    public float convertMilliliterToLiter(int milliliter) {
+    public float convertMilliliterToLiter(float milliliter) {
         System.out.println();
         float liter = milliliter / 1000.0f;
         return liter;
